+++ @udir-design/react/src/hooks/alpha.ts
@@ -6,10 +6,3 @@
  */
 
 export { useDrilldownTable } from './useDrilldownTable';
-export {
-  type FormNavigationState,
-  type GetFormNavigationStepProps,
-  type UseFormNavigationProps,
-  type UseFormNavigationReturn,
-  useFormNavigation,
-} from './useFormNavigation';
