+++ @udir-design/react/src/hooks/beta.ts
@@ -5,6 +5,13 @@
  * @module
  */
 
+export {
+  type FormNavigationState,
+  type GetFormNavigationStepProps,
+  type UseFormNavigationProps,
+  type UseFormNavigationReturn,
+  useFormNavigation,
+} from './useFormNavigation';
 export { type UsePaginationProps, usePagination } from './usePagination';
 export {
   useTableOfContents,
